// src/app/enums/fine-price.enum.ts
export enum FinePrice {
  FINE_20_CHF = 'FINE_20_CHF',
  FINE_40_CHF = 'FINE_40_CHF',
  FINE_60_CHF = 'FINE_60_CHF',
  FINE_100_CHF = 'FINE_100_CHF',
  FINE_120_CHF = 'FINE_120_CHF',
  FINE_160_CHF = 'FINE_160_CHF',
  FINE_180_CHF = 'FINE_180_CHF',
  FINE_240_CHF = 'FINE_240_CHF',
  FINE_250_CHF = 'FINE_250_CHF',
  FINE_260_CHF = 'FINE_260_CHF',
  POLICE_REPORT = 'POLICE_REPORT',
}